# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nuclom API",
    "version": "1.0.0",
    "description": "Nuclom is a video collaboration platform that helps teams organize, share, and collaborate on video content. This API provides programmatic access to videos, organizations, series, comments, and more.",
    "contact": {
      "name": "Nuclom Support",
      "url": "https://nuclom.com/support"
    },
    "license": {
      "name": "Elastic License 2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    }
  },
  "servers": [
    {
      "url": "https://nuclom.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/channels/{id}": {
      "delete": {
        "tags": [
          "Channels"
        ],
        "summary": "Delete channel",
        "operationId": "deleteChannel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "Get channel",
        "operationId": "getChannel",
        "description": "Get channel details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Channels"
        ],
        "summary": "Patch channel",
        "operationId": "patchChannel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/videos/{id}": {
      "delete": {
        "tags": [
          "Videos"
        ],
        "summary": "Delete video",
        "operationId": "deleteVideo",
        "description": "Delete a video. By default performs soft-delete with 30-day retention. Use permanent=true for immediate deletion.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "deletedAt": {
                          "type": "string"
                        },
                        "retentionUntil": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Video not found"
          }
        }
      },
      "get": {
        "tags": [
          "Videos"
        ],
        "summary": "Get video",
        "operationId": "getVideo",
        "description": "Get detailed information about a specific video including comments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video details",
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "UUID": {
                      "type": "string",
                      "description": "a Universally Unique Identifier",
                      "format": "uuid",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                    }
                  },
                  "type": "object",
                  "required": [
                    "id",
                    "title",
                    "description",
                    "duration",
                    "thumbnailUrl",
                    "videoUrl",
                    "authorId",
                    "organizationId",
                    "channelId",
                    "collectionId",
                    "transcript",
                    "aiSummary",
                    "viewCount",
                    "isPublic",
                    "isDeleted",
                    "deletedAt",
                    "retentionUntil",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "$ref": "#/$defs/UUID"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "duration": {
                      "type": "string"
                    },
                    "thumbnailUrl": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "videoUrl": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "authorId": {
                      "$ref": "#/$defs/UUID"
                    },
                    "organizationId": {
                      "$ref": "#/$defs/UUID"
                    },
                    "channelId": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/UUID"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "collectionId": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/UUID"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "transcript": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "aiSummary": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "viewCount": {
                      "type": "number"
                    },
                    "isPublic": {
                      "type": "boolean"
                    },
                    "isDeleted": {
                      "type": "boolean"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "retentionUntil": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "author": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "image"
                      ],
                      "properties": {
                        "id": {
                          "$ref": "#/$defs/UUID"
                        },
                        "name": {
                          "type": "string"
                        },
                        "image": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "additionalProperties": false
                    },
                    "comments": {
                      "type": "array",
                      "items": {
                        "$id": "/schemas/unknown",
                        "title": "unknown"
                      }
                    },
                    "chapters": {
                      "type": "array",
                      "items": {
                        "$id": "/schemas/unknown",
                        "title": "unknown"
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Video not found"
          }
        }
      },
      "put": {
        "tags": [
          "Videos"
        ],
        "summary": "Update video",
        "operationId": "updateVideo",
        "description": "Update video metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video updated",
            "content": {
              "application/json": {
                "schema": {
                  "$defs": {
                    "UUID": {
                      "type": "string",
                      "description": "a Universally Unique Identifier",
                      "format": "uuid",
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                    }
                  },
                  "type": "object",
                  "required": [
                    "id",
                    "title",
                    "description",
                    "duration",
                    "thumbnailUrl",
                    "videoUrl",
                    "authorId",
                    "organizationId",
                    "channelId",
                    "collectionId",
                    "transcript",
                    "aiSummary",
                    "viewCount",
                    "isPublic",
                    "isDeleted",
                    "deletedAt",
                    "retentionUntil",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "$ref": "#/$defs/UUID"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "duration": {
                      "type": "string"
                    },
                    "thumbnailUrl": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "videoUrl": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "authorId": {
                      "$ref": "#/$defs/UUID"
                    },
                    "organizationId": {
                      "$ref": "#/$defs/UUID"
                    },
                    "channelId": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/UUID"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "collectionId": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/UUID"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "transcript": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "aiSummary": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "viewCount": {
                      "type": "number"
                    },
                    "isPublic": {
                      "type": "boolean"
                    },
                    "isDeleted": {
                      "type": "boolean"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "retentionUntil": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Video not found"
          }
        }
      }
    },
...
    "/videos/{id}/move": {
      "post": {
        "tags": [
          "Videos"
        ],
        "summary": "Create move",
        "operationId": "createVideoMove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ID",
            "required": true,
            "schema": {
              "$defs": {
                "UUID": {
                  "type": "string",
                  "description": "a Universally Unique Identifier",
                  "format": "uuid",
                  "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                }
              },
              "$ref": "#/$defs/UUID"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication token from login or API key"
      }
    }
  }
}